move `ensure_genesis_block()` to `get_conn()`
